requirements.txt
alembic==1.8.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.1
Flask-Admin==1.6.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
python-dotenv==0.20.0
SQLAlchemy==1.4.39
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
